requirements.txt
uvicorn
fastapi[standard]
typing
pydantic
mysql-connector-python

# pymongo
# dnspython
# mysqlclient
